+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changed
     - Disallow replacing sample sheets if active landing zones exist (#713)
     - Display moved and deleted zones of other users with ``view_zones_all`` perm (#806)
     - Return landing zone iRODS path on creation (#843)
+    - Use ``SODARUserSerializer`` in ``LandingZoneSerializer`` (#842)
 - **Samplesheets**
     - Upgrade non-breaking Vue app dependencies (#836)
     - Reorganize views and URL patterns (#801)